[NUI] Load Tizen.NUI dependencies libs on Windows and MacOS
\### Description of Change ###
This creates a condition `NOTIZEN` at `build/common.props` than when
true defines `NOTIZEN`, an `MSBuild` compiler directive, which selects
dll names to use when compiling, thus enabling `Tizen.NUI` to correctly
search for libs at Mac and Windows.

This impacts the load of:
- `dali-csharp-binder`
- `capi-appfw-app-manager`
- `capi-appfw-app-common`
- `dlog`

We first tried to define those names at run-time, but
`DllImportAtribbuts` are `C#` `const strings`, which need to be defined
at compile-time, so it raised the need for the preprocessor directive to
select the correct name.

\### API Changes ###
Not exactly at API, but there were changes at `build.sh`:
- Use - instead of / to pass msbuild directives;
- As BSD readlink and expr are different than the one on Linux, this PR
  workaround those functions at build.sh, creating platform-independent
  functions that run at BSD and Linux Bash as well as Windows Git Bash,
  while maintaining the desired output on both systems;